fix ia32 build error. With IA32 tool chain, we can not use a 64bit data to divide a 32 bit data. MS tool chain may introduce an intrinsic function.
git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk/edk2@10268 6f19259b-4bc3-4df7-8a09-765794883524

+++ IntelFrameworkModulePkg/Bus/Pci/IdeBusDxe/Ata.c
@@ -2563,8 +2563,8 @@ AtaEnableLongPhysicalSector (
       //
       if ((AtaIdentifyData->alignment_logic_in_phy_blocks & 0xc000) == 0x4000) {
         IdeDev->BlkIo.Media->LowestAlignedLba =
-          (EFI_LBA) (IdeDev->BlkIo.Media->LogicalBlocksPerPhysicalBlock - (AtaIdentifyData->alignment_logic_in_phy_blocks & 0x3fff)) %
-          IdeDev->BlkIo.Media->LogicalBlocksPerPhysicalBlock;
+          (EFI_LBA) ((IdeDev->BlkIo.Media->LogicalBlocksPerPhysicalBlock - ((UINT32)AtaIdentifyData->alignment_logic_in_phy_blocks & 0x3fff)) %
+          IdeDev->BlkIo.Media->LogicalBlocksPerPhysicalBlock);
       }
     }
     //
